Add update-notifier
When the github command line tool is run, the `update-notifier` library
will check with the npm registry to see if a new version of this package
has been published, if it has been then it will display a notification
prompting the user to upgrade this package.

Note: This only applies when this package has been installed globally
with `npm` i.e. `npm install --global @financial-times/github`